+++ app/src/main/java/ly/count/android/demo/ActivityExampleRemoteConfig.java
@@ -84,6 +84,12 @@ public class ActivityExampleRemoteConfig extends AppCompatActivity {
         Countly.sharedInstance().remoteConfig().clearAll();
     }
 
+    public void onClickRemoteConfigEnrollForKeys(View v) {
+        String[] keys = { "aa", "bb" };
+        Countly.sharedInstance().remoteConfig().enrollIntoABTestsForKeys(keys);
+        Toast.makeText(getApplicationContext(), "Enrolling for Keys", Toast.LENGTH_SHORT).show();
+    }
+
     public void onClickRemoteConfigPrintValues(View v) {
         //this sample assumes that there are 5 keys available on the server